refactor: remove as much vertx as possible
