+++ .dir-locals.el
@@ -2,7 +2,7 @@
 ;;; For more information see (info "(emacs) Directory Variables")
 
 ((nil . (
-   (compile-command . "mvn -f /opt/minh/tools/gis/pom.xml -T 1C clean package")
+   (compile-command . "mvn -f ~/projects/gis/pom.xml -T 1C clean package")
    (projectile-project-compilation-cmd . "mvn -T 1C clean package")
-   (projectile-project-run-cmd . "cd /home/minh/projects/cm7/mem-system/src; java -jar /opt/minh/tools/gis/target/gis-1.1.4-dev.jar ")
+   (projectile-project-run-cmd . "cd /home/minh/projects/test/small-git-root-module; java -jar ~/projects/gis/target/gis-1.1.4-dev.jar ")
 )))
