+++ src/main/java/org/nqm/command/CommandVerticle.java
@@ -1,4 +1,4 @@
-package org.nqm.vertx;
+package org.nqm.command;
 
 import static org.nqm.command.GitCommand.HOOKS_OPTION;
 import static org.nqm.utils.GisStringUtils.isNotBlank;
@@ -18,12 +18,10 @@ import java.util.Optional;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.nqm.command.GitCommand;
 import org.nqm.config.GisConfig;
 import org.nqm.config.GisLog;
 import org.nqm.utils.GisStringUtils;
 import org.nqm.utils.StdOutUtils;
-import io.vertx.core.Future;
 
 public class CommandVerticle {
 
@@ -40,14 +38,6 @@ public class CommandVerticle {
     this.commandHook = extractHookCommand(args);
     this.commandWithArgs = buildCommandWithArgs(args);
     GisLog.debug("executing command '%s' under module '%s'", commandWithArgs, path.getFileName());
-    GisVertx.eventAddDir(path);
-  }
-
-  public CommandVerticle() {
-    this.path = null;
-    this.commandWithArgs = null;
-    this.commandHook = null;
-    GisVertx.eventAddDir(Path.of("."));
   }
 
   /**
@@ -66,13 +56,13 @@ public class CommandVerticle {
    */
   private String[] buildCommandWithArgs(String... args) {
     this.gisOptions = Stream.of(args)
-      .filter(arg -> arg.startsWith(OPTION_PREFIX))
-      .toArray(String[]::new);
+        .filter(arg -> arg.startsWith(OPTION_PREFIX))
+        .toArray(String[]::new);
 
     return Stream.concat(
-      Stream.of(GisConfig.GIT_HOME_DIR),
-      Stream.of(args).takeWhile(arg -> !arg.startsWith(OPTION_PREFIX) && !HOOKS_OPTION.equals(arg)))
-      .toArray(String[]::new);
+        Stream.of(GisConfig.GIT_HOME_DIR),
+        Stream.of(args).takeWhile(arg -> !arg.startsWith(OPTION_PREFIX) && !HOOKS_OPTION.equals(arg)))
+        .toArray(String[]::new);
   }
 
   public void execute() {
@@ -89,14 +79,14 @@ public class CommandVerticle {
 
     if (GisStringUtils.isNotBlank(this.commandHook)) {
       gisExecuteCommand(res, this.commandHook)
-          .forEach(f -> f.onComplete(r -> Optional.ofNullable(r.result()).ifPresent(p -> {
+          .forEach(p -> {
             try {
               infof("%s", new String(p.getInputStream().readAllBytes()));
             } catch (IOException e) {
               errln(e.getMessage());
               GisLog.debug(e);
             }
-          })));
+          });
     } else if (Stream.of(gisOptions).anyMatch("--gis-no-print-modules-name"::equals)) {
       safelyPrintWithoutModules(res);
     } else if (Stream.of(gisOptions).anyMatch("--gis-concat-modules-name"::equals)) {
@@ -105,35 +95,6 @@ public class CommandVerticle {
       safelyPrint(res);
     }
   }
-
-  // public void start() {
-  //   if (path == null) {
-  //     GisVertx.eventRemoveDir(Path.of("."));
-  //     return;
-  //   }
-  //   vertx.executeBlocking(() -> new ProcessBuilder(commandWithArgs).directory(path.toFile()).start(), false)
-  //       .compose(res -> {
-  //         if (GisStringUtils.isNotBlank(this.commandHook)) {
-  //           gisExecuteCommand(res, this.commandHook)
-  //               .forEach(f -> f.onComplete(r -> Optional.ofNullable(r.result()).ifPresent(p -> {
-  //                 try {
-  //                   infof("%s", new String(p.getInputStream().readAllBytes()));
-  //                 } catch (IOException e) {
-  //                   errln(e.getMessage());
-  //                   GisLog.debug(e);
-  //                 }
-  //               })));
-  //         } else if (Stream.of(gisOptions).anyMatch("--gis-no-print-modules-name"::equals)) {
-  //           safelyPrintWithoutModules(res);
-  //         } else if (Stream.of(gisOptions).anyMatch("--gis-concat-modules-name"::equals)) {
-  //           safelyConcatModuleNames(res);
-  //         } else {
-  //           safelyPrint(res);
-  //         }
-  //         return Future.succeededFuture();
-  //       })
-  //       .onComplete(r -> GisVertx.eventRemoveDir(path));
-  // }
 
   private void safelyPrint(Process pr) {
     var line = "";
@@ -145,24 +106,21 @@ public class CommandVerticle {
       while (isNotBlank(line = input.readLine())) {
         if (isStatusCmd) {
           sb.append(isOneLineOpt ? gitStatusOneLine(line) : gitStatus(line));
-        }
-        else {
+        } else {
           sb.append("%n  %s".formatted(line));
         }
       }
       StdOutUtils.println(sb.toString());
       Optional.of(pr.waitFor())
-        .filter(exitCode -> exitCode != 0)
-        .ifPresent(exitCode -> {
-          GisLog.debug(EXIT_WITH_CODE_MSG_FMT.formatted(exitCode));
-          warnln(WARN_MSG_FMT.formatted(this.path.getFileName()));
-        });
-    }
-    catch (IOException e) {
+          .filter(exitCode -> exitCode != 0)
+          .ifPresent(exitCode -> {
+            GisLog.debug(EXIT_WITH_CODE_MSG_FMT.formatted(exitCode));
+            warnln(WARN_MSG_FMT.formatted(this.path.getFileName()));
+          });
+    } catch (IOException e) {
       errln(e.getMessage());
       GisLog.debug(e);
-    }
-    catch (InterruptedException e) {
+    } catch (InterruptedException e) {
       GisLog.debug(e);
       Thread.currentThread().interrupt();
     }
@@ -178,35 +136,31 @@ public class CommandVerticle {
       }
       StdOutUtils.print(sb.toString());
       Optional.of(pr.waitFor())
-        .filter(exitCode -> exitCode != 0)
-        .ifPresent(exitCode -> {
-          GisLog.debug(EXIT_WITH_CODE_MSG_FMT.formatted(exitCode));
-          warnln(WARN_MSG_FMT.formatted(this.path.getFileName()));
-        });
-    }
-    catch (IOException e) {
+          .filter(exitCode -> exitCode != 0)
+          .ifPresent(exitCode -> {
+            GisLog.debug(EXIT_WITH_CODE_MSG_FMT.formatted(exitCode));
+            warnln(WARN_MSG_FMT.formatted(this.path.getFileName()));
+          });
+    } catch (IOException e) {
       errln(e.getMessage());
       GisLog.debug(e);
-    }
-    catch (InterruptedException e) {
+    } catch (InterruptedException e) {
       GisLog.debug(e);
       Thread.currentThread().interrupt();
     }
   }
 
-  private List<Future<Process>> gisExecuteCommand(Process pr, String cmd) {
+  private List<Process> gisExecuteCommand(Process pr, String cmd) {
     var input = new BufferedReader(new InputStreamReader(pr.getInputStream()));
-    var futures = new ArrayList<Future<Process>>();
+    var futures = new ArrayList<Process>();
     try {
       var line = "";
       while (isNotBlank(line = input.readLine())) {
-        futures.add(Future.succeededFuture(
-          new ProcessBuilder(cmd.formatted(line).split(" "))
+        futures.add(new ProcessBuilder(cmd.formatted(line).split(" "))
             .directory(path.toFile())
-            .start()));
+            .start());
       }
-    }
-    catch (IOException e) {
+    } catch (IOException e) {
       errln(e.getMessage());
       GisLog.debug(e);
     }
