+++ src/main/java/org/nqm/command/Wrapper.java
@@ -1,19 +1,19 @@
 package org.nqm.command;
 
 import static org.nqm.config.GisConfig.CURRENT_DIR;
-import static org.nqm.utils.StdOutUtils.errln;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 import org.nqm.config.GisLog;
 import org.nqm.exception.GisException;
-import org.nqm.vertx.CommandVerticle;
+import org.nqm.utils.StdOutUtils;
 
 public final class Wrapper {
 
@@ -21,11 +21,6 @@ public final class Wrapper {
 
   private static File isFileExist(File f) {
     return f.exists() ? f : null;
-  }
-
-  public static void deployVertx(Path path, String... args) {
-    getFileMarker();
-    new CommandVerticle(path, args).execute();
   }
 
   private static File getFileMarker() {
@@ -48,6 +43,9 @@ public final class Wrapper {
   public static void forEachModuleWith(Predicate<Path> pred, String... args) {
     var gitModulesFilePath = getFileMarker();
     try (var exe = Executors.newVirtualThreadPerTaskExecutor()) {
+      Optional.of(Path.of(CURRENT_DIR))
+          .filter(pred)
+          .ifPresent(root -> exe.submit(() -> new CommandVerticle(root, args).execute()));
       Files.readAllLines(gitModulesFilePath.toPath()).stream()
           .map(String::trim)
           .filter(s -> s.startsWith("path"))
@@ -57,13 +55,13 @@ public final class Wrapper {
             if (dir.toFile().exists()) {
               return true;
             }
-            System.err.println("directory '%s' does not exist, will be ignored!".formatted("" + dir));
+            StdOutUtils.errln("directory '%s' does not exist, will be ignored!".formatted("" + dir));
             return false;
           })
           .filter(pred)
           .forEach(dir -> exe.submit(() -> new CommandVerticle(dir, args).execute()));
     } catch (IOException e) {
-      errln("failed to read file '.gis-modules' or '.gitmodules'");
+      StdOutUtils.errln("failed to read file '.gis-modules' or '.gitmodules'");
     }
   }
